requirements.txt
# Core dependencies for weather display system
Pillow>=9.0.0
cairosvg>=2.7.0

# Data processing for preview system
pandas>=2.0.0

# CircuitPython compatibility layer for desktop development
adafruit-blinka>=8.22.0
adafruit-platformdetect>=3.70.1

# Display and graphics libraries (CircuitPython compatible)
adafruit-circuitpython-displayio>=4.0.0
adafruit-circuitpython-display-text>=2.26.4
adafruit-circuitpython-bitmap-font>=2.3.2
adafruit-circuitpython-display-shapes>=2.7.0

# HTTP and networking
adafruit-circuitpython-requests>=1.14.0
adafruit-circuitpython-connection-manager>=1.0.0
requests>=2.28.0

# Utilities
adafruit-circuitpython-datetime>=1.6.11

# Hardware libraries (for desktop simulation/testing)
adafruit-circuitpython-sdcard>=3.3.9
adafruit-circuitpython-ssd1683>=1.2.0
adafruit-circuitpython-hdc302x>=1.0.2

# Image processing and bitmap utilities
adafruit-circuitpython-bitmapsaver>=1.3.9
adafruit-circuitpython-imageload>=1.24.5

# Development dependencies (optional)
# For enhanced font handling and image processing
# fonttools>=4.42.0        # For font handling
# pygame>=2.5.0            # For display simulation
blinka-displayio-pygamedisplay @ git+https://github.com/FoamyGuy/Blinka_Displayio_PyGameDisplay.git  # For desktop DisplayIO preview

# Note: CircuitPython libraries are automatically available on the microcontroller
# and don't need to be installed via pip when deploying to hardware.
#
# For development/preview on desktop:
# pip install -r requirements.txt
#
# For microcontroller deployment:
# Just copy the 300x400/CIRCUITPY/ directory to the device
